* Generators should support large particle numbers (not too sure about droplet generator) * Progress information for cubic grid generator
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@1290 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ tools/gui/generators/CubicGridGenerator.h
@@ -31,7 +31,7 @@
 class CubicGridGenerator: public MDGenerator {
 
 private:
-	int _numMolecules;
+	long int _numMolecules;
 	double _molarDensity;
 	std::vector<Component> _components;
 	double _temperature;
